+++ drain_eye/lib/data/repositories_impl/inspection_repository_impl.dart
@@ -1,150 +1,2 @@
-import 'dart:convert';
-import 'dart:io';
-
-import 'package:drain_eye/data/datasources/local_damage_model_datasource.dart';
-import 'package:drain_eye/data/models/inspection_model.dart';
-import 'package:drain_eye/domain/entities/inspection.dart';
-import 'package:drain_eye/domain/entities/model_inference_result.dart';
-import 'package:drain_eye/domain/repositories/inspection_repository.dart';
-import 'package:flutter/foundation.dart';
-import 'package:http/http.dart' as http;
-import 'package:http_parser/http_parser.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-
-// реализация репозитория: GET инспекций с сервера, tflite, POST новой инспекции
-class InspectionRepositoryImpl implements InspectionRepository {
-  InspectionRepositoryImpl({
-    http.Client? httpClient,
-    LocalDamageModelDataSource? damageModelDataSource,
-    SharedPreferences? prefs,
-  })  : _httpClient = httpClient ?? http.Client(),
-        _damageModel = damageModelDataSource ?? LocalDamageModelDataSource(),
-        _prefs = prefs;
-
-  final http.Client _httpClient;
-  final LocalDamageModelDataSource _damageModel;
-  final SharedPreferences? _prefs;
-
-  // ссылка будет меняться
-  final String baseUrl = 'https://9pnoz7-95-161-60-178.ru.tuna.am';
-  static const _authTokenPrefsKey = 'auth_token';
-
-  @override
-  Stream<List<Inspection>> getUserInspections() {
-    return Stream.fromFuture(_fetchMyInspections());
-  }
-
-  /// GET `/inspector/my_inspections` — список инспекций текущего инженера.
-  Future<List<Inspection>> _fetchMyInspections() async {
-    final token = await _authToken();
-    final uri = Uri.parse('$baseUrl/inspector/my_inspections');
-    final response = await _httpClient.get(
-      uri,
-      headers: {
-        'Accept': 'application/json',
-        'Authorization': 'Bearer $token',
-      },
-    );
-    if (response.statusCode < 200 || response.statusCode >= 300) {
-      throw Exception(
-        'my_inspections: ${response.statusCode} ${response.body}',
-      );
-    }
-    final decoded = jsonDecode(response.body);
-    if (decoded is! List) {
-      throw FormatException('Ожидался JSON-массив инспекций');
-    }
-    final out = <Inspection>[];
-    for (var i = 0; i < decoded.length; i++) {
-      final raw = decoded[i];
-      if (raw is! Map) {
-        throw FormatException('Элемент списка инспекций не объект');
-      }
-      final dto = InspectionModel.fromJson(Map<String, dynamic>.from(raw));
-      out.add(dto.toDomain(index: i));
-    }
-    if (kDebugMode) {
-      print('>>> my_inspections: ${out.length} записей');
-    }
-    return out;
-  }
-
-  @override
-  Future<ModelInferenceResult> runDamageModel(List<String> photoPaths) {
-    return _damageModel.runAverageProbabilities(photoPaths);
-  }
-
-  @override
-  Future<void> createInspection({
-    required int userId,
-    required List<String> photoPaths,
-    required ModelInferenceResult modelResult,
-  }) async {
-    final token = await _authToken();
-    final uri = Uri.parse('$baseUrl/inspector/add_inspection');
-
-
-    // FastAPI: inspection_json: str = Form(...), files: List[UploadFile] = File(None)
-    // ModelVerdictSchema — все поля обязательны; null из приложения заменяем заглушками.
-    final inspectionPayload = <String, dynamic>{
-      'engineer_id': userId.toString(),
-      'timestamp': DateTime.now().toIso8601String(),
-      'model_verdict': <String, dynamic>{
-        'material': modelResult.material ?? 'unknown',
-        'state': modelResult.state ?? 0,
-        'damage_type': modelResult.damageType,
-        'damage_degree': modelResult.damageDegree ?? 0.0,
-        'accuracy_model': modelResult.accuracyModel,
-        'comments': modelResult.comments ?? '',
-      },
-      'address': '',
-      'name': 'Инспекция',
-      'photos': <String>[],
-      'status_sync': 'save',
-    };
-
-    final request = http.MultipartRequest('POST', uri)
-      ..headers['Accept'] = 'application/json'
-      ..headers['Authorization'] = 'Bearer $token'
-      ..fields['inspection_json'] = jsonEncode(inspectionPayload);
-
-    for (final p in photoPaths) {
-      final f = File(p);
-      if (!await f.exists()) {
-        throw StateError('Файл снимка не найден: $p');
-      }
-      request.files.add(
-        await http.MultipartFile.fromPath(
-          'files',
-          p,
-          filename: _basename(p),
-          contentType: MediaType('image', 'jpeg'),
-        ),
-      );
-    }
-
-    final streamed = await _httpClient.send(request);
-    final response = await http.Response.fromStream(streamed);
-    if (response.statusCode < 200 || response.statusCode >= 300) {
-      throw Exception(
-        'add_inspection: ${response.statusCode} ${response.body}',
-      );
-    }
-  }
-
-  static String _basename(String path) {
-    final i = path.lastIndexOf('/');
-    final j = path.lastIndexOf(r'\');
-    final k = i > j ? i : j;
-    return k < 0 ? path : path.substring(k + 1);
-  }
-
-  Future<String> _authToken() async {
-    final prefs = _prefs ?? await SharedPreferences.getInstance();
-    final token = prefs.getString(_authTokenPrefsKey);
-    if (token == null || token.isEmpty) {
-      throw StateError('токен авторизации не найден');
-    }
-    return token;
-  }
-}
+export 'inspection_repository_impl_native.dart'
+    if (dart.library.html) 'inspection_repository_impl_web.dart';
